refactor: split raid commands across multiple files for organization
